+++ ecs/util.go
@@ -87,6 +87,11 @@ func (p *pagedSlice[T]) Get(index int32) *T {
 	return &p.pages[index/pageSize][index%pageSize]
 }
 
+// Set sets the value at the given index.
+func (p *pagedSlice[T]) Set(index int32, value T) {
+	p.pages[index/pageSize][index%pageSize] = value
+}
+
 // Len returns the current number of items in the paged slice.
 func (p *pagedSlice[T]) Len() int32 {
 	return p.len